requirements.txt
streamlit>=1.35.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.7.0
python-dotenv>=1.0.0
pandas>=2.0.0
undetected-chromedriver>=3.5.5
